+++ src/main/java/gift/controller/IndexController.java
@@ -1,22 +1,34 @@
 package gift.controller;
 
+import gift.config.KakaoProperties;
 import gift.dto.CategoryDTO;
 import gift.dto.ProductDTO;
 import gift.service.CategoryService;
+import gift.service.KakaoAuthService;
 import gift.service.ProductService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.util.UriComponentsBuilder;
 
 import java.util.ArrayList;
 import java.util.List;
 
 @Controller
 public class IndexController {
+    private static final Logger logger = LoggerFactory.getLogger(KakaoController.class);
+
+    private final KakaoAuthService kakaoAuthService;
+    private final KakaoProperties kakaoProperties;
+
     private final ProductService productService;
     private final CategoryService categoryService;
 
-    public IndexController(ProductService productService, CategoryService categoryService) {
+    public IndexController(KakaoAuthService kakaoAuthService, KakaoProperties kakaoProperties, ProductService productService, CategoryService categoryService) {
+        this.kakaoAuthService = kakaoAuthService;
+        this.kakaoProperties = kakaoProperties;
         this.productService = productService;
         this.categoryService = categoryService;
     }
@@ -51,5 +63,17 @@ public class IndexController {
         categoryList.addAll(GetCategoryList);
         categoryList.remove("NONE");
         return categoryList;
+    }
+
+    @GetMapping(value="/oauth/kakao")
+    public String kakaoConnect() {
+        logger.info("redirect_uri={}", kakaoProperties.getRedirectUrl());
+        logger.info("client_id=" + kakaoProperties.getClientId());
+        String url = UriComponentsBuilder.fromHttpUrl(kakaoProperties.getLoginUrl())
+                .queryParam("redirect_uri", kakaoProperties.getRedirectUrl())
+                .queryParam("client_id", kakaoProperties.getClientId())
+                .build()
+                .toUriString();
+        return "redirect:" + url;
     }
 }
