+++ src/main/resources/application.properties
@@ -10,5 +10,5 @@ spring.sql.init.mode = always
 spring.jpa.hibernate.ddl-auto=update
 
 kakao.client-id=4d60db641a23118784b0e087bf586cbc
-kakao.redirect-url=http://localhost:8080/kakaoAuth
+kakao.redirect-url=http://localhost:8080/oauth/kakao/callback
 kakao.login-url=https://kauth.kakao.com/oauth/authorize?scope=talk_message&response_type=code&
