+++ src/main/resources/templates/index.html
@@ -43,7 +43,7 @@
     Admin
 </h1>
 <div>
-    <form th:action="@{/kakao}" method="Get" style="display: inline;">
+    <form th:action="@{/oauth/kakao}" method="Get" style="display: inline;">
         <button type="submit">KakaoLogin</button>
     </form><br>
     <form th:action="@{/postform}" method="Get" style="display: inline;">
